Handheld & Windows
1. GR Handheld 	- Delivery Note & bill of Lading always leave blank
2. GR Handheld 	- Qty - if qty exceeded, reason pop out
3. FG Handheld 	- Remove Quantity textbox
		- Remove Qty Received at datagrid
4. FG Windows 	- Label - Rename Section to Total Section
		- Change to QR Code

+++ Windows_Application/DataLayer/Repositories/GoodsReceiveRepository.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using System.Collections.Generic;
 using System.Data.Entity;
+using System;
 
 namespace DataLayer.Repositories
 {
@@ -75,14 +76,20 @@ namespace DataLayer.Repositories
         }
 
         public void Delete(long ID)
-        {
-            using (var context = new InventoryContext())
+        { using (var context = new InventoryContext())
             {
+
+                var GITxn = context.GITransactions.Where(id => id.GRID == ID);
+                if (GITxn.Any())
+                    context.GITransactions.RemoveRange(GITxn);
+
                 var GRTxn = context.GRTransactions.Where(id => id.GRID == ID);
-                context.GRTransactions.RemoveRange(GRTxn);
+                if (GRTxn.Any())
+                    context.GRTransactions.RemoveRange(GRTxn);
 
                 var EUNTxn = context.EunKGs.Where(id => id.GRID == ID);
-                context.EunKGs.RemoveRange(EUNTxn);
+                if (EUNTxn.Any())
+                    context.EunKGs.RemoveRange(EUNTxn);
 
                 var GR = context.GoodsReceives.Where(id => id.ID == ID);
                 context.GoodsReceives.RemoveRange(GR);
